Configure automapper for models and change some values for db models - team stats byte to int

+++ StatScore/StatScore.Data/Models/LeagueStats.cs
@@ -17,10 +17,10 @@
 
         public virtual League League { get; init; }
 
-        public byte Wins { get; set; }
+        public int Wins { get; set; }
 
-        public byte Losses { get; set; }
+        public int Losses { get; set; }
 
-        public byte Draws { get; set; }
+        public int Draws { get; set; }
     }
 }
